feat: :rocket: Added a container for the header title, included the synopsis key in data objects, adjusted the "View More" button and search functionality to handle promises, and modified styles with media queries.
A container for the header title was added to improve the management of the element. The synopsis key was included in the data objects to display this information in the pop-up cards of the films. The "View More" button functionality was modified to adapt a promise in one of its logic parts. Additionally, search functionalities in methods were adjusted to incorporate promises for better asynchronous operation handling. Finally, styles of the webpage were modified, and respective media queries were added to ensure proper display across different screen sizes.